requirements.txt
discord.py
requests
beautifulsoup4
python-dotenv